+++ components/Checkbox/src/Checkbox.tsx
@@ -1,5 +1,5 @@
 import { CheckIcon } from '@hautechai/webui.icon';
-import { styled } from '@linaria/react';
+import { styled } from '@hautechai/webui.themeprovider';
 import { themeVars } from '@hautechai/webui.themeprovider';
 import { PropsWithChildren } from 'react';