+++ components/FloatingPanel/src/FloatingPanel.tsx
@@ -1,7 +1,7 @@
 import { ArrowAltDownIcon, ArrowAltUpIcon } from '@hautechai/webui.icon';
 import { IconButton } from '@hautechai/webui.iconbutton';
 import { Progress } from '@hautechai/webui.progress';
-import { styled } from '@linaria/react';
+import { styled } from '@hautechai/webui.themeprovider';
 import { themeVars } from '@hautechai/webui.themeprovider';
 import { Typography } from '@hautechai/webui.typography';