+++ components/MenuItem/src/MenuItem.tsx
@@ -1,5 +1,5 @@
 import { Row } from '@hautechai/webui.row';
-import { styled } from '@linaria/react';
+import { styled } from '@hautechai/webui.themeprovider';
 import { themeVars } from '@hautechai/webui.themeprovider';
 import { Typography } from '@hautechai/webui.typography';
 import React from 'react';