+++ components/VisualEditorInput/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.4](https://github.com/HautechAI/webui/compare/@hautechai/webui.visualeditorinput@2.1.3...@hautechai/webui.visualeditorinput@2.1.4) (2025-08-17)
+
+### Bug Fixes
+
+- Revert linaria back to emotion ([#102](https://github.com/HautechAI/webui/issues/102)) ([e229790](https://github.com/HautechAI/webui/commit/e229790dae8eba4b3037bbe41365e5a73ab7f6dc))
+
+## [2.1.3](https://github.com/HautechAI/webui/compare/@hautechai/webui.visualeditorinput@2.1.2...@hautechai/webui.visualeditorinput@2.1.3) (2025-08-16)
+
+**Note:** Version bump only for package @hautechai/webui.visualeditorinput
+
 ## [2.1.2](https://github.com/HautechAI/webui/compare/@hautechai/webui.visualeditorinput@2.1.1...@hautechai/webui.visualeditorinput@2.1.2) (2025-08-16)
 
 **Note:** Version bump only for package @hautechai/webui.visualeditorinput